KA-67 Refactor file structure and names
